requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-limiter==3.7.0
python-dotenv==1.0.1
bcrypt==4.1.3
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-cov==5.0.0
